fix useEffect() infinite loop
eslint wants the functions we're using in our useEffect() to be part of the dependency array, but our functions are `() => {}` so it changes every time and loops. stupid to use the entire props in the dep array regardless, that's my bad. i believe this is a case where we'd want to use `useCallback()` but i'll get that working in a different PR